endpoint de creacion comentarios acomodado a un mismo archivo con el resto de los otros endpoint
