+++ controllers/comments.controller.js
@@ -1,4 +1,5 @@
 const db = require('../models/index');
+const { Comment } = require("../models");
 
 const getAllComments =  async (req,res) =>{
     Comment.findAll({attributes: ["body"], order: [["createdAt", "ASC"],]}).then(response =>   
@@ -8,8 +9,28 @@ const getAllComments =  async (req,res) =>{
         });
 }
 
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+const uploadComment = async (req, res) => {
+    const { userId, body, newsId } = req.body;
+  
+    try {
+      const comment = await Comment.create({
+          userId,
+          body,
+          newsId,
+        });
+      
+        res.status(200).send(comment)
+    } catch (error) {
+      res.status(404).send("ups something went wrong....try again")
+    }
+  };
+
 
 module.exports = {
     getAllComments,
+    uploadComment
   };
